debug

+++ app/models/schemas.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 
 class summarize_text(BaseModel):
     text: str
-    precent: int
+    precent: float
